third_party/WebKit/LayoutTests/imported/web-platform-tests/html/browsers/the-window-object/window-indexed-properties-expected.txt

This is a testharness.js-based test.
PASS Indexed properties of the window object (non-strict mode) 
FAIL Indexed properties of the window object (non-strict mode) 1 assert_equals: expected (object) object "[object Window]" but got (string) "foo"
FAIL Indexed properties of the window object (non-strict mode) 2 assert_equals: expected (undefined) undefined but got (string) "foo"
FAIL Indexed properties of the window object (non-strict mode) 3 assert_false: 0 in proto expected false got true
Harness: the test ran to completion.
